Create README.md
This project proposes to automate the process of map creation for the Washington State Wetland Rating System with an arcgis tool from a python script. 
Federal, state and local governments provide the basic data needed for map production.
The proposed modeling sets a workflow for multiple inputs with different geoprocessing tools and the intended work flow is meant to cover the entire wetlands rating process from data creation to the final map presentation.